+++ EasyReact/Classes/Core/EZRMutableNode.m
@@ -150,10 +150,25 @@ static inline EZSFliterBlock _EZR_PropertyExists(NSString *keyPath) {
 }
 
 - (void)setValue:(id)value context:(nullable id)context {
+    [self setValue:value senderList:nil context:context];
+}
+
+- (void)setValue:(id)value senderList:(nullable EZRSenderList *)senderList context:(nullable id)context {
     if EZR_LikelyNO(!self.isMutable) {
         EZR_THROW(EZRNodeExceptionName, EZRExceptionReason_CannotModifyEZRNode, nil);
     }
-    [self next:value from:[EZRSenderList new] context:context];
+#ifdef DEBUG
+    if EZR_LikelyNO(!(self.hasListener || self.hasDownstreamNode)) {
+        NSLog(@"[EasyReact Warning] The node has no listeners .The value is: %@",value);
+    }
+#endif
+    EZRSenderList *currentSenderList = nil;
+    if (senderList == nil) {
+        currentSenderList = [EZRSenderList new];
+    } else {
+        currentSenderList = senderList;
+    }
+    [self next:value from:currentSenderList context:context];
 }
 
 - (void)clean {