+++ EasyReact/Classes/Core/NodeTransforms/EZRCaseTransform.m
@@ -63,14 +63,9 @@ static BOOL EZR_instanceEqual(id left, id right) {
         self.lastNode = next.second;
         @ezr_weakify(self)
         self.cancelable = [[next.second listenedBy:self] withSenderListAndContextBlock:^(id  _Nullable value, EZRSenderList * _Nonnull insideSenderList, id  _Nullable insideContext) {
-           @ezr_strongify(self)
-            if (self.cancelable) {
-                [self _superNext:value from:insideSenderList context:insideContext];
-            }
+            @ezr_strongify(self)
+            [self _superNext:value from:insideSenderList context:insideContext];
         }];
-        if (!next.second.isEmpty) {
-            [super next:next.second.value from:[senderList appendNewSender:next.second] context:context];
-        }
     }
 }